requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
drf_spectacular==0.26.5
mailjet_rest==1.3.4
python-dotenv==1.0.1
django-debug-toolbar==3.2.4
psycopg2-binary==2.9.9